Refactor: investments에서 Comments로 수정 + 투자지원모달에서 지원자모달로 변경
