+++ app/company-detail/header/Description.tsx
