+++ app/company-detail/header/Header.tsx
