+++ app/company-detail/header/Stats.tsx
